refactor: decompose MarkdownEditor.tsx into focused modules (phase 4)
